Update main.cpp
Instead of using magic numbers for sensor types, I defined them so people new to the codebase can see which sensors are which.

Also, when you define ARDNAME as "garage", every time you use it you create a brand new character array. Instead, I defined the character array as a constant so we don't create a copy every time.